Refactor DataReducer
