+++ public/index.html
@@ -7,7 +7,7 @@
   <meta name="theme-color" content="#000000">
   <link rel="stylesheet" href="https://gitchui.netlify.app/style.css">
   <title>NFTbaazar</title>
-  <link rel="shortcut icon" href="../src/resources/title.gif">
+  <link rel="icon" href="./title.gif" type="image/gif" />
 </head>
 
 <body>
