+++ src/Contexts/index.js
@@ -1,2 +1,2 @@
-export { useData, DataProvider } from "./DataContext";
+export { useData, DataProvider } from "./data/DataContext";
 export { useAuth, AuthProvider } from "./AuthContext";
